+++ io_packages/kh2lib.lua
@@ -6,7 +6,7 @@ KH2_LIBRARY_VERSION = 2
 
 -- This represents the "full" version including any potential hotfix-type changes that wouldn't need to increment the
 -- major version. It's meant for display purposes only to help with debugging/troubleshooting.
-KH2_LIBRARY_DISPLAY_VERSION = "2.0.0"
+KH2_LIBRARY_DISPLAY_VERSION = "2.1"
 
 -- Constants for each game version - scripts can check this using `GameVersion`
 KH2_VERSION_UNKNOWN = 0x0000
@@ -74,31 +74,29 @@ function ReadPointer(address)
     end
 end
 
--- Logs a message to the console, using an appropriate call per platform.
+-- Logs a message to the console
 function Log(message)
-    if kh2lib.OnPC then
-        ConsolePrint(message)
-    else
-        print(message)
-    end
+    ConsolePrint(tostring(message))
 end
 
--- Logs a warning message to the console, using an appropriate call per platform.
+-- Logs a message to the console, with "MESSAGE: " prefix
+function LogMessage(message)
+    ConsolePrint(tostring(message), 0)
+end
+
+-- Logs a success message to the console
+function LogSuccess(message)
+    ConsolePrint(tostring(message), 1)
+end
+
+-- Logs a warning message to the console
 function LogWarning(message)
-    if kh2lib.OnPC then
-        ConsolePrint(message, 2)
-    else
-        print("WARNING: " .. message)
-    end
+    ConsolePrint(tostring(message), 2)
 end
 
--- Logs an error message to the console, using an appropriate call per platform.
+-- Logs an error message to the console
 function LogError(message)
-    if kh2lib.OnPC then
-        ConsolePrint(message, 3)
-    else
-        print("ERROR: " .. message)
-    end
+    ConsolePrint(tostring(message), 3)
 end
 
 local function _AddInventoryAddresses(table)
@@ -424,6 +422,13 @@ local function _InitLibrary()
             kh2lib.OnPC = true
             kh2lib.CanExecute = true
         elseif ReadByte(0x660EF4) == 106 then
+            -- Steam Global
+            kh2lib = require("kh2lib.SteamShared-1_0_0_10")
+            kh2lib.GameVersion = KH2_VERSION_STEAM_1_0_0_10
+            kh2lib.OnPC = true
+            kh2lib.CanExecute = true
+        elseif ReadByte(0x660E74) == 106 then
+            -- Steam JP (different address for detection but the other addresses are shared)
             kh2lib = require("kh2lib.SteamShared-1_0_0_10")
             kh2lib.GameVersion = KH2_VERSION_STEAM_1_0_0_10
             kh2lib.OnPC = true
@@ -432,13 +437,13 @@ local function _InitLibrary()
             kh2lib.GameVersion = KH2_VERSION_UNKNOWN
             kh2lib.OnPC = true
             kh2lib.CanExecute = false
-            LogError(baseVersionMessage .. "KH2 PC detected, but is a version not currently supported")
+            LogError("KH2 Lua Library - KH2 PC detected, but is a version not currently supported")
         end
     else
         kh2lib.GameVersion = KH2_VERSION_UNKNOWN
         kh2lib.OnPC = false
         kh2lib.CanExecute = false
-        LogError(baseVersionMessage .. "KH2 not detected")
+        LogError("KH2 Lua Library - KH2 not detected")
     end
 
     if kh2lib.CanExecute then